refactor: move `~/middleware` dir support back to pages
