--- examples/with-universal-router/middleware/always-run.global.ts
@@ -1,5 +1,0 @@
-export default defineNuxtRouteMiddleware((to) => {
-  if (to.path === '/forbidden') {
-    return false
-  }
-})
